pyproject.toml
[project]
name = "mcat-gen"
version = "0.1.0"
description = "Generate MCAT practice questions using a local LLM"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "vllm>=0.6.0",
]

[project.scripts]
mcat-gen = "src.main:main"
